Move types to dashboard path
