+++ client/components/sites-add-new-site/async.tsx
@@ -1,6 +1,6 @@
 import { Spinner } from '@wordpress/components';
 import AsyncLoad from 'calypso/components/async-load';
-import type { AddNewSiteProps } from './types';
+import type { AddNewSiteProps } from '../../dashboard/sites/add-new-site/types';
 
 export const AsyncContent = ( props: AddNewSiteProps ) => {
 	return (
