+++ client/components/sites-add-new-site/index.tsx
@@ -3,7 +3,7 @@ import { Dropdown } from '@wordpress/components';
 import { useTranslate } from 'i18n-calypso';
 import { AsyncContent } from './async';
 import AddNewSiteButton from './button';
-import type { AddNewSiteProps } from './types';
+import type { AddNewSiteProps } from '../../dashboard/sites/add-new-site/types';
 import type { PropsWithChildren } from 'react';
 import './style.scss';
