+++ client/components/sites-add-new-site/test/index.tsx
@@ -23,22 +23,24 @@ jest.mock( '../async', () => ( {
 } ) );
 
 describe( 'SitesAddNewSitePopover', () => {
+	const sitesDashboardContext = 'sites-dashboard';
+
 	beforeEach( () => {
 		jest.clearAllMocks();
 	} );
 
 	it( 'renders with full label when not in compact mode', () => {
-		render( <SitesAddNewSitePopover showCompact={ false } /> );
+		render( <SitesAddNewSitePopover showCompact={ false } context={ sitesDashboardContext } /> );
 		expect( screen.getByText( 'Add new site' ) ).toBeInTheDocument();
 	} );
 
 	it( 'renders without label in compact mode', () => {
-		render( <SitesAddNewSitePopover showCompact /> );
+		render( <SitesAddNewSitePopover showCompact context={ sitesDashboardContext } /> );
 		expect( screen.queryByText( 'Add new site' ) ).not.toBeInTheDocument();
 	} );
 
 	it( 'tracks event when dropdown is opened', () => {
-		render( <SitesAddNewSitePopover showCompact={ false } /> );
+		render( <SitesAddNewSitePopover showCompact={ false } context={ sitesDashboardContext } /> );
 		const button = screen.getByRole( 'button' );
 		fireEvent.click( button );
 		expect( recordTracksEvent ).toHaveBeenCalledWith(
@@ -47,7 +49,7 @@ describe( 'SitesAddNewSitePopover', () => {
 	} );
 
 	it( 'closes popover when clicking the button again', async () => {
-		render( <SitesAddNewSitePopover showCompact={ false } /> );
+		render( <SitesAddNewSitePopover showCompact={ false } context={ sitesDashboardContext } /> );
 		const button = screen.getByRole( 'button' );
 		fireEvent.click( button );
 		expect( screen.getByTestId( 'async-content' ) ).toBeInTheDocument();
